requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
pillow==11.3.0
sqlparse==0.5.3
